chore: distinct pending and accepted quote pages
